requirements.txt
Django==1.4.20
MySQL-python==1.2.5
django-annoying==0.8.3
django-multifilefield==1.0
django-settings-export==1.0.5
djangorestframework==3.1.3
python-dateutil==2.4.2
python-ldap==2.4.20
